+++ src/utils.h
@@ -8,13 +8,13 @@ namespace SCAMP {
 void elementwise_max_with_index(std::vector<float> &mp_full,
                                 std::vector<uint32_t> &mpi_full,
                                 int64_t merge_start, int64_t tile_sz,
-                                std::vector<unsigned long long> *to_merge);
+                                std::vector<uint64_t> *to_merge);
 
 void compute_statistics(const double *T, double *norms, double *df, double *dg,
                         double *means, size_t n, size_t m, cudaStream_t s,
                         double *scratch);
 
-void launch_merge_mp_idx(float *mp, uint32_t *mpi, uint32_t n,
-                         unsigned long long *merged, cudaStream_t s);
+void launch_merge_mp_idx(float *mp, uint32_t *mpi, uint32_t n, uint64_t *merged,
+                         cudaStream_t s);
 
 }  // namespace SCAMP
